requirements.txt
alembic==0.9.9
asn1crypto==0.24.0
bcrypt==3.1.4
blinker==1.4
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.3
dominate==2.3.1
elasticsearch==6.3.1
enum34==1.1.6
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Dance==0.14.0
Flask-DebugToolbar==0.10.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Moment==0.6.0
Flask-OAuth==0.12
Flask-OAuthlib==0.9.4
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.0
Flask-WTF==0.14.2
gunicorn==19.9.0
httplib2==0.11.3
idna==2.6
ipaddress==1.0.17
itsdangerous==0.24
Jinja2==2.10
keyring==10.6.0
keyrings.alt==3.0
lazy==1.3
Mako==1.0.7
MarkupSafe==1.0
numpy==1.15.1
oauth2==1.9.0.post1
oauthlib==2.0.7
pathlib==1.0.1
Pillow==5.3.0
psycopg2==2.7.5
pycparser==2.18
pycrypto==2.6.1
pygobject==3.26.1
PyMySQL==0.9.2
python-dateutil==2.7.2
python-dotenv==0.9.1
python-editor==1.0.3
pyxdg==0.25
pyzbar==0.1.7
qrcode==6.0
reportlab==3.5.9
requests==2.18.4
requests-oauthlib==0.8.0
SecretStorage==2.3.1
six==1.11.0
SQLAlchemy==1.2.6
urllib3==1.22
URLObject==2.4.3
virtualenv==16.1.0
visitor==0.1.3
Werkzeug==0.14.1
Whoosh==2.6.0
WTForms==2.1
